sandboxgame/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sandboxgame"
version = "0.1.0"
description = "Two-storey house defense sandbox rendered with PyOpenGL."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Sandbox Game Team" }
]
dependencies = [
    "PyOpenGL>=3.1.6",
    "glfw>=2.5.0",
]

[project.optional-dependencies]
tests = [
    "pytest>=7.4",
]

[project.urls]
Homepage = "https://example.com/sandboxgame"
Repository = "https://example.com/sandboxgame/repo"

[tool.pytest.ini_options]
testpaths = ["tests"]
